% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary.R
\docType{data}
\name{peccary}
\alias{peccary}
\title{Time of the day when a peccary was observed}
\format{
A vector of 16 observations (in radians).
}
\usage{
data(peccary)
}
\description{
Temporal activity information (time of the day in radians)
when a camera detected the appearance of a peccary at El Triunfo
biosphere in Mexico in 2015 data provided by Eduardo Mendoza from
\emph{Universidad Michoacana de San Nicolas de Hidalgo}, Mexico.
}
\references{
Nieto-Barajas, L.E. & Nunez-Antonio, G. (2019). Projected Polya tree. https://arxiv.org/pdf/1902.06020.pdf
}
\keyword{datasets}
