% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorppt.summary.R
\name{priorppt.summary}
\alias{priorppt.summary}
\title{Summary for the prior projected Polya tree simulations}
\usage{
priorppt.summary(priorppt.circ, units = "radians")
}
\arguments{
\item{priorppt.circ}{object returned by  \code{dsimpriorppt} function.}

\item{units}{units of the support: "radians", "degrees" or "hours".}
}
\value{
Table of descriptive statistics for mean direction and concentration.
}
\description{
Mean, quantiles 2.5\% and 97.5\% of the mean direction and concentration.
}
\examples{
z <- dsimpriorppt(mu = c(-1,0), nsim = 5, units = "hours")
priorppt.summary(z)

}
