% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapir.R
\docType{data}
\name{tapir}
\alias{tapir}
\title{Time of the day when a tapir was observed}
\format{
A vector of 35 observations (in radians).
}
\usage{
data(tapir)
}
\description{
Temporal activity information (time of the day in radians)
when a camera detected the appearance of a tapir at El Triunfo
biosphere in Mexico in 2015 data provided by Eduardo Mendoza from
\emph{Universidad Michoacana de San Nicolas de Hidalgo}, Mexico.
}
\references{
Nieto-Barajas, L.E. & Nunez-Antonio, G. (2019). Projected Polya tree. https://arxiv.org/pdf/1902.06020.pdf
}
\keyword{datasets}
